20 June 2022 Java OOPS Polymorphism , Interface and Abstraction

+++ src/test/java/com/java/oops/overLoadingbyDifferentDataTypesWithDifferentArgumentTypes.java
@@ -1,7 +1,7 @@
 package com.java.oops;
 
 class Adder123 {
-	static int add(int a, int b) {
+	static double add(int a, double b) {
 		return a + b;
 	}
